+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/importers/svg/css/CssSymbolType.java
@@ -1,0 +1,36 @@
+package com.jpexs.decompiler.flash.importers.svg.css;
+
+/**
+ *
+ * @author JPEXS
+ */
+public enum CssSymbolType {
+    S,
+    CDO,
+    CDC,
+    INCLUDES,
+    DASHMATCH,
+    STRING,
+    BAD_STRING,
+    IDENT,
+    HASH,
+    IMPORT_SYM,
+    PAGE_SYM,
+    MEDIA_SYM,
+    CHARSET_SYM,
+    IMPORTANT_SYM,
+    EMS,
+    EXS,
+    LENGTH,
+    ANGLE,
+    TIME,
+    FREQ,
+    DIMENSION,
+    PERCENTAGE,
+    NUMBER,
+    URI,
+    BAD_URI,
+    FUNCTION,
+    EOF,
+    OTHER
+}